Aggregate cases in epichains objects according to a grouping variable

Aggregate cases in epichains objects according to a grouping variable

Usage

# S3 method for epichains
aggregate(x, grouping_var = c("time", "generation", "both"), ...)

Arguments

x

An epichains object.

grouping_var

The variable to group and count over. Options include "time", "generation", and "both".

...

Other arguments passed to aggregate.

Value

If grouping_var is either "time" or "generation", a data.frame with cases aggregated over grouping_var; If grouping_var = "both", a list of data.frames, the first being for cases over time, and the second being for cases over generations.

Author

James M. Azam

Examples

set.seed(123)
chains <- simulate_tree(
  nchains = 10, statistic = "size",
  offspring_dist = "pois", stat_max = 10, serials_sampler = function(x) 3,
  lambda = 2
)
chains
#> `epichains` object
#> 
#> < tree head (from first known ancestor) >
#> 
#>    chain_id sim_id ancestor generation time
#> 11        1      2        1          2    3
#> 12        2      2        1          2    3
#> 15        3      2        1          2    3
#> 17        4      2        1          2    3
#> 21        5      2        1          2    3
#> 25        7      2        1          2    3
#> 
#> < tree tail >
#> 
#>     chain_id sim_id ancestor generation time
#> 125       10     17        8          4    9
#> 70         5     18        5          3    6
#> 126       10     18        8          4    9
#> 71         5     19        5          3    6
#> 127       10     19        8          4    9
#> 72         5     20        5          3    6
#> Chains simulated: 10
#> Number of ancestors (known): 8
#> Number of generations: 4
#> Use `as.data.frame(<object_name>)` to view the full output in the console.

# Aggregate cases per time
aggregate(chains, grouping_var = "time")
#>   time cases
#> 1    0    10
#> 2    3    24
#> 3    6    64
#> 4    9    29

# Aggregate cases per generation
aggregate(chains, grouping_var = "generation")
#>   generation cases
#> 1          1    10
#> 2          2    24
#> 3          3    64
#> 4          4    29

# Aggregate cases per both time and generation
aggregate(chains, grouping_var = "both")
#> [[1]]
#>   time cases
#> 1    0    10
#> 2    3    24
#> 3    6    64
#> 4    9    29
#> 
#> [[2]]
#>   generation cases
#> 1          1    10
#> 2          2    24
#> 3          3    64
#> 4          4    29
#> 
#> attr(,"class")
#> [1] "epichains_aggregate_df" "list"                  
#> attr(,"chain_type")
#> [1] "chains_tree"
#> attr(,"aggregated_over")
#> [1] "both"