Log-likelihood of the summary (size/length) of chains with generic offspring distribution

The log-likelihoods are calculated with a crude approximation using simulated chain summaries by linearly approximating any missing values in the empirical cumulative distribution function (ecdf).

Usage

offspring_ll(x, offspring_dist, statistic, nsim_offspring = 100, ...)

Arguments

x

Vector of chain summaries (sizes/lengths)

offspring_dist

Offspring distribution: a <character> string corresponding to the R distribution function (e.g., "pois" for Poisson, where rpois is the R function to generate Poisson random numbers).

statistic

<String>; Chain statistic to track as the stopping criteria for each chain being simulated when stat_max is not Inf. Can be one of:

"size": the total number of cases produced by a chain before it goes extinct.

"length": the total number of ancestors produced by a chain before it goes extinct.

nsim_offspring

Number of simulations of the offspring distribution for approximating the distribution of the chain statistic summary (size/length)

...

any parameters to pass to simulate_summary

Value

log-likelihood values

See also

Author

Sebastian Funk

Examples

set.seed(123)
chain_size_ll <- offspring_ll(
  x = c(1, 5, 6, 8, 7, 8, 10),
  offspring_dist = "pois",
  statistic = "size",
  lambda = 0.82
)