Generate random numbers from the Borel distribution

Random numbers are generated by simulating from a Poisson branching process

Usage

rborel(n, mu, censor_at = Inf)

Arguments

n

Number of random variates to generate.

mu

A non-negative number for the poisson mean.

censor_at

A stopping criterion; <numeric>. Defaults to Inf. A value above which the simulation ends and the random number is set to Inf (as a form of censoring). rborel() simulates chain sizes using simulate_chain_stats() with a Poisson offspring distribution, so if mu >= 1, the simulation could proceed unendingly. This parameter is used to prevent this.

Value

A numeric vector of random numbers.

Author

Sebastian Funk

Examples

set.seed(32)
rborel(5, 1)
#> [1]  25  50 114   5   1