Simulate transmission chains sizes/lengths

Simulate transmission chains sizes/lengths

Usage

simulate_summary(
  ntrees,
  statistic = c("size", "length"),
  offspring_dist,
  stat_max = Inf,
  ...
)

Arguments

ntrees

Number of trees to simulate.

statistic

String; Statistic (size/length) to calculate. Used to determine stopping criteria for simulations when stat_max is finite. Can be one of:

"size": the total number of offspring.

"length": the total number of infectors.

offspring_dist

Offspring distribution: a character string corresponding to the R distribution function (e.g., "pois" for Poisson, where rpois is the R function to generate Poisson random numbers).

stat_max

A cut off for the chain statistic (size/length) being computed. Results above the specified value, are set to Inf.

...

Parameters of the offspring distribution as required by R.

Calculating chain sizes and lengths

The function simulates the chain size for individual \(i\) at time \(t\), \(I_{t, i}\), as: $$I_{t, i} = \sum_{i}^{I_{t-1}}X_{t, i},$$ and the chain length/duration for individual \(i\) at time \(t\), \(L_{t, i}\), as: $$L_{t, i} = {\sf min}(1, X_{t, i}), $$ where \(X_{t, i}\) is the secondary cases generated by individual \(i\) at time \(t\), and \(I_{0, i} = L_{0, i} = 1\).

The distribution of secondary cases, \(X_{t, i}\) is modelled by the offspring distribution (offspring_dist).

Specifying generation_time

generation_time must be specified as a function with one argument.

For example, assuming we want to specify the generation time as a random log-normally distributed variable with meanlog = 0.58 and sdlog = 1.58, we could define a named function, let's call it "generation_time_fn", with only one argument representing the number of generation times to sample: generation_time_fn <- function(n){rlnorm(n, 0.58, 1.38)}, and assign the name of the function to generation_time in the simulation function, i.e. `simulate_*`(..., generation_time = generation_time_fn), where ... are the other arguments to simulate_*() and * is a placeholder for the rest of simulation function's name.

Alternatively, we could assign an anonymous function to generation_time in the simulate_*() call, i.e. simulate_*(..., generation_time = function(n){rlnorm(n, 0.58, 1.38)}), where ... are the other arguments to simulate_*().

See also

Author

James M. Azam, Sebastian Funk

Examples

simulate_summary(
  ntrees = 10,
  statistic = "size",
  offspring_dist = "pois",
  stat_max = 10,
  lambda = 2
)
#> `epichains_summary` object 
#> 
#>  [1] Inf Inf Inf Inf Inf Inf Inf   1 Inf Inf
#> 
#>  Simulated tree sizes: 
#> 
#> Min: 1