Simulate transmission trees from an initial number of infections

Simulate transmission trees from an initial number of infections

Usage

simulate_tree(
  nchains,
  statistic = c("size", "length"),
  offspring_dist,
  stat_max = Inf,
  intvn_mean_reduction = 0,
  serials_dist,
  t0 = 0,
  tf = Inf,
  ...
)

Arguments

nchains

Number of chains to simulate.

statistic

String; Statistic (size/length) to calculate. Used to determine stopping criteria for simulations when stat_max is finite. Can be one of:

"size": the total number of offspring.

"length": the total number of ancestors.

offspring_dist

Offspring distribution: a character string corresponding to the R distribution function (e.g., "pois" for Poisson, where rpois is the R function to generate Poisson random numbers).

stat_max

A cut off for the chain statistic (size/length) being computed. Results above the specified value, are set to this value. Defaults to Inf.

intvn_mean_reduction

A number between 0 and 1 for scaling/reducing the mean of offspring_dist. Serves as population-level intervention. intvn_mean_reduction = 0 implies no intervention impact and intvn_mean_reduction = 1 implies full impact.

serials_dist

The serial interval distribution function; the name of a user-defined named or anonymous function with only one argument n, representing the number of serial intervals to generate. See details.

t0

Start time (if serial interval is given); either a single value or a vector of same length as nchains (number of simulations) with initial times. Defaults to 0.

tf

End time (if serial interval is given).

...

Parameters of the offspring distribution as required by R.

Value

An <epichains> object, which is basically a <data.frame> with columns chain_id (chain ID), sim_id (a unique ID within each simulation for each individual), ancestor

(the ID of the ancestor of each individual), generation, and time (of infection)

Calculating chain sizes and lengths

The function simulates the chain size for individual \(i\) at time \(t\), \(I_{t, i}\), as: $$I_{t, i} = \sum_{i}^{I_{t-1}}X_{t, i},$$ and the chain length/duration for individual \(i\) at time \(t\), \(L_{t, i}\), as: $$L_{t, i} = {\sf min}(1, X_{t, i}), $$ where \(X_{t, i}\) is the secondary cases generated by individual \(i\) at time \(t\), and \(I_{0, i} = L_{0, i} = 1\).

The distribution of secondary cases, \(X_{t, i}\) is modelled by the offspring distribution (offspring_dist).

The serial interval (serials_dist)

Assumptions/disambiguation

In epidemiology, the generation interval is the duration between successive infectious events in a chain of transmission. Similarly, the serial interval is the duration between observed symptom onset times between successive cases in a transmission chain. The generation interval is often hard to observe because exact times of infection are hard to measure hence, the serial interval is often used instead . Here, we use the serial interval to represent what would normally be called the generation interval, that is, the time between successive cases.

See References below for some literature on the subject.

Specifying serials_dist

serials_dist must be specified as a named or anonymous/inline/unnamed function with one argument.

For example, assuming we want to specify the serial interval distribution as a random log-normally distributed variable with meanlog = 0.58 and sdlog = 1.58, we could define a named function, let's call it "serial_interval", with only one argument representing the number of serial intervals to sample: serial_interval <- function(n){rlnorm(n, 0.58, 1.38)}, and assign the name of the function to serials_dist in the simulation function, i.e. `simulate_*`(..., serials_dist = serial_interval), where ... are the other arguments to simulate_*() and * is a placeholder for the rest of simulation function's name.

Alternatively, we could assign an anonymous function to serials_dist in the simulate_*() call, i.e. simulate_*(..., serials_dist = function(n){rlnorm(n, 0.58, 1.38)}), where ... are the other arguments to simulate_*().

References

Lehtinen S, Ashcroft P, Bonhoeffer S. On the relationship between serial interval, infectiousness profile and generation time. J R Soc Interface. 2021 Jan;18(174):20200756. doi:10.1098/rsif.2020.0756 . Epub 2021 Jan 6. PMID: 33402022; PMCID: PMC7879757.

Fine PE. The interval between successive cases of an infectious disease. Am J Epidemiol. 2003 Dec 1;158(11):1039-47. doi:10.1093/aje/kwg251. PMID: 14630599

Jacob C. (2010). Branching processes: their role in epidemiology. International journal of environmental research and public health, 7(3), 1186–1204. doi:10.3390/ijerph7031204

See also

Author

James M. Azam, Sebastian Funk

Examples

set.seed(123)
chains <- simulate_tree(
  nchains = 10,
  statistic = "size",
  offspring_dist = "pois",
  stat_max = 10,
  serials_dist = function(x) 3,
  lambda = 2
)

# Run model with intervention a 50% reduction in R0.
chains_with_intvn <- simulate_tree(
  nchains = 10,
  statistic = "size",
  offspring_dist = "pois",
  intvn_mean_reduction = 0.5,
  stat_max = 10,
  serials_dist = function(x) 3,
  lambda = 2
)

chains_with_intvn
#> `epichains` object
#> 
#> < tree head (from first known ancestor) >
#> 
#>    chain_id sim_id ancestor generation time
#> 11        1      2        1          2    3
#> 12        4      2        1          2    3
#> 14        5      2        1          2    3
#> 16        6      2        1          2    3
#> 17        7      2        1          2    3
#> 18        9      2        1          2    3
#> 
#> < tree tail >
#> 
#>    chain_id sim_id ancestor generation time
#> 42        4     10        7          5   12
#> 48        5     10        8          5   12
#> 43        4     11        7          5   12
#> 44        4     12        7          5   12
#> 45        4     13        8          5   12
#> 46        4     14        8          5   12
#> Chains simulated: 10
#> Number of ancestors (known): 8
#> Number of generations: 6
#> Use `as.data.frame(<object_name>)` to view the full output in the console.