Simulate a tree of infections from an initial susceptible population with initial immunity

Simulate a tree of infections from an initial susceptible population with initial immunity

Usage

simulate_tree_from_pop(
  pop,
  offspring_dist = c("pois", "nbinom"),
  offspring_mean,
  offspring_disp,
  serials_dist,
  initial_immune = 0,
  t0 = 0,
  tf = Inf
)

Arguments

pop

The susceptible population.

offspring_dist

Offspring distribution: a character string corresponding to the R distribution function (e.g., "pois" for Poisson, where rpois is the R function to generate Poisson random numbers).

offspring_mean

The average number of secondary cases for each case. Same as R0.

offspring_disp

The dispersion parameter of the number of secondary cases. Ignored if offspring == "pois". Must be > 1 to avoid division by 0 when calculating the size. See details and ?rnbinom for details on the parameterisation in Ecology.

serials_dist

The serial interval. A function that takes one parameter (n), the number of serial intervals to randomly sample. Value must be >= 0.

initial_immune

The number of initial immunes in the population.

t0

Start time; Defaults to 0.

tf

End time; Defaults to Inf.

Value

a data frame with columns time, id (a unique ID for each individual element of the chain), ancestor (the ID of the ancestor of each element), and generation.

Offspring distributions

Currently only "pois" & "nbinom" are supported. Internally truncated distributions are used to avoid infecting more people than susceptibles available.

The poisson model is parametrised so that:

lamda = offspring_mean * pop - initial_immune / pop

The negative binomial model is parametrised as:

mu = offspring_mean * pop - initial immune / pop, and size = mu / (offspring_disp - 1). This is why offspring_disp must be greater than 1.

simulate_tree_from_pop() has a couple of key different from simulate_tree():

the maximal chain statistic is limited by pop instead of stat_max (in simulate_tree()),

it can only handle implemented offspring distributions ("pois" and "nbinom").

Author

Flavio Finger

James M. Azam

Examples

# Simulate with poisson offspring
simulate_tree_from_pop(
  pop = 100, offspring_dist = "pois",
  offspring_mean = 0.5, serials_dist = function(x) 3
)
#> `epichains` object
#> 
#> < tree head (from first known ancestor) >
#> 
#> [1] sim_id     ancestor   generation time      
#> <0 rows> (or 0-length row.names)
#> 
#> < tree tail >
#> 
#>   sim_id ancestor generation time
#> 1      1       NA          1    0
#> Number of ancestors (known): 0
#> Number of generations: 1
#> Use `as.data.frame(<object_name>)` to view the full output in the console.

# Simulate with negative binomial offspring
simulate_tree_from_pop(
  pop = 100, offspring_dist = "nbinom",
  offspring_mean = 0.5, offspring_disp = 1.1, serials_dist = function(x) 3
)
#> `epichains` object
#> 
#> < tree head (from first known ancestor) >
#> 
#> [1] sim_id     ancestor   generation time      
#> <0 rows> (or 0-length row.names)
#> 
#> < tree tail >
#> 
#>   sim_id ancestor generation time
#> 1      1       NA          1    0
#> Number of ancestors (known): 0
#> Number of generations: 1
#> Use `as.data.frame(<object_name>)` to view the full output in the console.