Estimate the log-likelihood/likelihood for observed branching processes

Estimate the log-likelihood/likelihood for observed branching processes

Usage

likelihood(
  chains,
  statistic = c("size", "length"),
  offspring_dist,
  nsim_obs,
  log = TRUE,
  obs_prob = 1,
  stat_max = Inf,
  exclude = NULL,
  individual = FALSE,
  ...
)

Arguments

chains

Vector of chain summaries (sizes/lengths)

statistic

String; Statistic (size/length) to calculate. Used to determine stopping criteria for simulations when stat_max is finite. Can be one of:

"size": the total number of offspring.

"length": the total number of ancestors.

offspring_dist

Offspring distribution: a character string corresponding to the R distribution function (e.g., "pois" for Poisson, where rpois is the R function to generate Poisson random numbers).

nsim_obs

Number of simulations if the log-likelihood/likelihood is to be approximated for imperfect observations.

log

Logical; Should the log-likelihoods be transformed to likelihoods? (Defaults to TRUE).

obs_prob

Observation probability (assumed constant)

stat_max

A cut off for the chain statistic (size/length) being computed. Results above the specified value, are set to Inf.

exclude

A vector of indices of the sizes/lengths to exclude from the log-likelihood calculation.

individual

If TRUE, a vector of individual log-likelihood/likelihood contributions will be returned rather than the sum/product.

...

any parameters to pass to simulate_summary

Value

If log = TRUE

A joint log-likelihood (sum of individual log-likelihoods), if individual == FALSE (default) and obs_prob == 1 (default), or

A list of individual log-likelihoods, if individual == TRUE and obs_prob == 1 (default), or

A list of individual log-likelihoods (same length as nsim_obs), if individual == TRUE and 0 <= obs_prob < 1, or

A vector of joint log-likelihoods (same length as nsim_obs), if individual == FALSE and 0 <= obs_prob < 1 (imperfect observation).

If log = FALSE, the same structure of outputs as above are returned, except that likelihoods, instead of log-likelihoods, are calculated in all cases. Moreover, the joint likelihoods are the product, instead of the sum, of the individual likelihoods.

See also

Author

Sebastian Funk

Examples

# example of observed chain sizes
set.seed(121)
# randomly generate 20 chains of size 1 to 10
chain_sizes <- sample(1:10, 20, replace = TRUE)
likelihood(
  chains = chain_sizes, statistic = "size",
  offspring_dist = "pois", nsim_obs = 100, lambda = 0.5
)
#> [1] -67.82879