Simulate a tree of infections with a serial and offspring distributions

Simulate a tree of infections with a serial and offspring distributions

simulate_tree(
  nchains,
  statistic = c("size", "length"),
  offspring_dist,
  stat_max = Inf,
  serials_sampler,
  t0 = 0,
  tf = Inf,
  ...
)

Arguments

nchains

Number of chains to simulate.

statistic

String; Statistic to calculate. Can be one of:

"size": the total number of offspring.

"length": the total number of ancestors.

offspring_dist

Offspring distribution: a character string corresponding to the R distribution function (e.g., "pois" for Poisson, where rpois is the R function to generate Poisson random numbers).

stat_max

A cut off for the chain statistic (size/length) being computed. Results above the specified value, are set to this value. Defaults to Inf.

serials_sampler

The serial interval generator function; the name of a user-defined named or anonymous function with only one argument n, representing the number of serial intervals to generate.

t0

Start time (if serial interval is given); either a single value or a vector of same length as nchains (number of simulations) with initial times. Defaults to 0.

tf

End time (if serial interval is given).

...

Parameters of the offspring distribution as required by R.

Value

an epichains object, which is basically a data.frame with columns chain_id (chain ID), sim_id (a unique ID within each simulation for each individual element of the chain), ancestor

(the ID of the ancestor of each element), generation, and time (of infection)

Details

simulate_tree() simulates a branching process of the form: WIP

The serial interval (serials_sampler):

Assumptions/disambiguation

In epidemiology, the generation interval is the duration between successive infectious events in a chain of transmission. Similarly, the serial interval is the duration between observed symptom onset times between successive cases in a transmission chain. The generation interval is often hard to observe because exact times of infection are hard to measure hence, the serial interval is often used instead . Here, we use the serial interval to represent what would normally be called the generation interval, that is, the time between successive cases.

See References below for some literature on the subject.

Specifying serials_sampler in simulate_tree()

serials_sampler must be specified as a named or anonymous/inline/unnamed function # nolint with one argument.

For example, assuming we want to specify the serial interval generator as a random log-normally distributed variable with meanlog = 0.58 and sdlog = 1.58, we could define a named function, let's call it "serial_interval", with only one argument representing the number of serial intervals to sample: serial_interval <- function(n){rlnorm(n, 0.58, 1.38)}, and assign the name of the function to serials_sampler in simulate_tree() like so simulate_tree(..., serials_sampler = serial_interval), where ... are the other arguments to simulate_tree().

Alternatively, we could assign an anonymous function to serials_sampler in the simulate_tree() call like so simulate_tree(..., serials_sampler = function(n){rlnorm(n, 0.58, 1.38)}), #nolint where ... are the other arguments to simulate_tree().

References

Lehtinen S, Ashcroft P, Bonhoeffer S. On the relationship between serial interval, infectiousness profile and generation time. J R Soc Interface. 2021 Jan;18(174):20200756. doi: 10.1098/rsif.2020.0756. Epub 2021 Jan 6. PMID: 33402022; PMCID: PMC7879757.

Fine PE. The interval between successive cases of an infectious disease. Am J Epidemiol. 2003 Dec 1;158(11):1039-47. doi: 10.1093/aje/kwg251. PMID: 14630599.

See also

Author

James M. Azam, Sebastian Funk

Examples

set.seed(123)
chains <- simulate_tree(
  nchains = 10, statistic = "size",
  offspring_dist = "pois", stat_max = 10, serials_sampler = function(x) 3,
  lambda = 2
)