Simulate transmission trees from a susceptible or partially immune population

Simulate transmission trees from a susceptible or partially immune population

Usage

simulate_tree_from_pop(
  pop,
  offspring_dist = c("pois", "nbinom"),
  serials_dist,
  initial_immune = 0,
  t0 = 0,
  tf = Inf,
  ...
)

Arguments

pop

The susceptible population size.

offspring_dist

Offspring distribution: a character string corresponding to the R distribution function (e.g., "pois" for Poisson, where rpois is the R function to generate Poisson random numbers). Only supports "pois" and "nbinom".

serials_dist

The serial interval distribution function; the name of a user-defined named or anonymous function with only one argument n, representing the number of serial intervals to generate. See details.

initial_immune

The number of initial immunes in the population. Must be less than pop - 1.

t0

Start time; Defaults to 0.

tf

End time; Defaults to Inf.

...

Parameters of the offspring distribution as required by R.

Value

An <epichains> object, which is basically a <data.frame> with columns sim_id (a unique ID within each simulation for each individual of the chain), ancestor (the ID of the ancestor of each individual), generation, and time (of infection).

Offspring distributions

Currently, offspring_dist only supports "pois" & "nbinom". Internally, the respective truncated poisson and negative binomial distributions are used to avoid the situation where there are more cases than susceptibles at any point.

The poisson model has mean, lambda, parametrised as: $${\sf lambda} = \dfrac{{\sf lambda} \times ({\sf pop} - {\sf initial\_immune} - 1)}{{\sf pop}}$$

The negative binomial model, has mean, mu, parametrised as: $${\sf mu} = \dfrac{{\sf mu} \times ({\sf pop} - {\sf initial\_immune} - 1)}{{\sf pop}},$$ and dispersion, size, parametrised as: $${\sf size} = \dfrac{{\sf mu}}{{\sf size} - 1}.$$ This is why size must be greater than 1.

Differences with simulate_tree()

simulate_tree_from_pop() has a couple of key differences from simulate_tree():

the maximal chain statistic is limited by pop instead of stat_max (in simulate_tree()),

offspring_dist can only handle "pois" and "nbinom".

The serial interval (serials_dist)

Assumptions/disambiguation

In epidemiology, the generation interval is the duration between successive infectious events in a chain of transmission. Similarly, the serial interval is the duration between observed symptom onset times between successive cases in a transmission chain. The generation interval is often hard to observe because exact times of infection are hard to measure hence, the serial interval is often used instead . Here, we use the serial interval to represent what would normally be called the generation interval, that is, the time between successive cases.

See References below for some literature on the subject.

Specifying serials_dist

serials_dist must be specified as a named or anonymous/inline/unnamed function with one argument.

For example, assuming we want to specify the serial interval distribution as a random log-normally distributed variable with meanlog = 0.58 and sdlog = 1.58, we could define a named function, let's call it "serial_interval", with only one argument representing the number of serial intervals to sample: serial_interval <- function(n){rlnorm(n, 0.58, 1.38)}, and assign the name of the function to serials_dist in the simulation function, i.e. `simulate_*`(..., serials_dist = serial_interval), where ... are the other arguments to simulate_*() and * is a placeholder for the rest of simulation function's name.

Alternatively, we could assign an anonymous function to serials_dist in the simulate_*() call, i.e. simulate_*(..., serials_dist = function(n){rlnorm(n, 0.58, 1.38)}), where ... are the other arguments to simulate_*().

See also

Author

Flavio Finger, James M. Azam, Sebastian Funk

Examples

# Simulate with poisson offspring
simulate_tree_from_pop(
  pop = 100,
  offspring_dist = "pois",
  lambda = 0.5,
  serials_dist = function(x) 3
)
#> `epichains` object
#> 
#> < tree head (from first known ancestor) >
#> 
#> [1] sim_id     ancestor   generation time      
#> <0 rows> (or 0-length row.names)
#> 
#> < tree tail >
#> 
#>   sim_id ancestor generation time
#> 1      1       NA          1    0
#> Number of ancestors (known): 0
#> Number of generations: 1
#> Use `as.data.frame(<object_name>)` to view the full output in the console.

# Simulate with negative binomial offspring
simulate_tree_from_pop(
pop = 100, offspring_dist = "nbinom",
mu = 0.5,
size = 1.1,
serials_dist = function(x) 3
)
#> `epichains` object
#> 
#> < tree head (from first known ancestor) >
#> 
#> [1] sim_id     ancestor   generation time      
#> <0 rows> (or 0-length row.names)
#> 
#> < tree tail >
#> 
#>   sim_id ancestor generation time
#> 1      1       NA          1    0
#> Number of ancestors (known): 0
#> Number of generations: 1
#> Use `as.data.frame(<object_name>)` to view the full output in the console.